S302T500-236 #time 20m Changed the tests to delete the businesses admins as well as the generated users and businesses ?testing

+++ backend/src/test/java/org/seng302/datagenerator/BusinessGeneratorTest.java
@@ -24,7 +24,7 @@ public class BusinessGeneratorTest {
         //Connects to production database
         String url = "jdbc:mariadb://localhost/seng302-2021-team500-prod";
         //change password
-        this.conn = DriverManager.getConnection(url, "seng302-team500", "ListenDirectly6053");
+        this.conn = DriverManager.getConnection(url, "seng302-team500", "changeMe");
 
         //Creates generators
         this.userGenerator = new UserGenerator(conn);
@@ -67,23 +67,39 @@ public class BusinessGeneratorTest {
         return results.getLong(1);
     }
 
-    public void deleteUsersFromDb(List<Long> userIds) throws SQLException {
-        for (Long userId: userIds) {
-            PreparedStatement stmt = conn.prepareStatement("DELETE FROM user WHERE userid = ?");
-            stmt.setObject(1, userId);
-            stmt.executeUpdate();
-        }
-    }
-
     /**
-     * Deletes the generated business from the database as part of the clean up process
+     * Deletes the generated users from the database as part of the clean up process
      * @param businessIds the ids of the generated businesses
+     * @param userIds the ids of the generated users
      */
-    public void deleteBusinessesFromDB(List<Long> businessIds) throws SQLException {
-        for (Long businessId: businessIds) {
-            PreparedStatement stmt = conn.prepareStatement("DELETE FROM business WHERE id = ?");
-            stmt.setObject(1, businessId);
+    public void deleteUsersAndBusinessesFromDb(List<Long> businessIds, List<Long> userIds) throws SQLException {
+        if (businessIds != null) {
+            long lowestUserId = userIds.get(0);
+            long highestUserId = userIds.get(userIds.size()-1);
+            long lowestBusinessId = businessIds.get(0);
+            long highestBusinessId = businessIds.get(businessIds.size()-1);
+
+            PreparedStatement stmt = conn.prepareStatement(
+                    "DELETE FROM business_admins WHERE user_id >= ? AND user_id <= ? AND " +
+                            "business_id >= ? AND business_id <= ?"
+            );
+            stmt.setObject(1, lowestUserId);
+            stmt.setObject(2, highestUserId);
+            stmt.setObject(3, lowestBusinessId);
+            stmt.setObject(4, highestBusinessId);
             stmt.executeUpdate();
+
+            for (Long businessId: businessIds) {
+                PreparedStatement stmt2 = conn.prepareStatement("DELETE FROM business WHERE id = ?");
+                stmt2.setObject(1, businessId);
+                stmt2.executeUpdate();
+            }
+        }
+
+        for (Long userId: userIds) {
+            PreparedStatement stmt3 = conn.prepareStatement("DELETE FROM user WHERE userid = ?");
+            stmt3.setObject(1, userId);
+            stmt3.executeUpdate();
         }
     }
 
@@ -97,8 +113,7 @@ public class BusinessGeneratorTest {
         long businessId = businessIds.get(0);
         checkRequiredFieldsNotNull(businessId);
         //clean up
-        deleteBusinessesFromDB(businessIds);
-        deleteUsersFromDb(userIds);
+        deleteUsersAndBusinessesFromDb(businessIds, userIds);
     }
 
     @Test
@@ -112,8 +127,7 @@ public class BusinessGeneratorTest {
             checkRequiredFieldsNotNull(businessIds.get(i));
         }
 
-        deleteBusinessesFromDB(businessIds);
-        deleteUsersFromDb(userIds);
+        deleteUsersAndBusinessesFromDb(businessIds, userIds);
     }
 
     @Test
@@ -127,8 +141,7 @@ public class BusinessGeneratorTest {
             checkRequiredFieldsNotNull(businessIds.get(i));
         }
 
-        deleteBusinessesFromDB(businessIds);
-        deleteUsersFromDb(userIds);
+        deleteUsersAndBusinessesFromDb(businessIds, userIds);
     }
 
     @Test
@@ -142,8 +155,7 @@ public class BusinessGeneratorTest {
             checkRequiredFieldsNotNull(businessIds.get(i));
         }
 
-        deleteBusinessesFromDB(businessIds);
-        deleteUsersFromDb(userIds);
+        deleteUsersAndBusinessesFromDb(businessIds, userIds);
     }
 
     @Test
@@ -156,7 +168,7 @@ public class BusinessGeneratorTest {
             fail();
         }
 
-        deleteUsersFromDb(userIds);
+        deleteUsersAndBusinessesFromDb(null, userIds);
     }
 
     @Test
@@ -169,7 +181,7 @@ public class BusinessGeneratorTest {
             fail();
         }
 
-        deleteUsersFromDb(userIds);
+        deleteUsersAndBusinessesFromDb(null, userIds);
     }
 
     @Test
@@ -182,6 +194,6 @@ public class BusinessGeneratorTest {
             fail();
         }
 
-        deleteUsersFromDb(userIds);
+        deleteUsersAndBusinessesFromDb(null, userIds);
     }
 }
